Hide methods of obtaining entity type info
